Allow completion repeats to continue repeating from following lines like the built-in completions do
The newline plus any indent is removed, and just the first word from the following line is matched. For this, the CompleteHelper#Repeat#Processor() is offered. It transforms the leading whitespace of the match, and manipulates the internal completion records so that the original completion text is kept (and therefore continues to match), even though only the word itself has been inserted as a completion candidate.
Allow to cancel repeats via CompleteHelper#Repeat#Clear().