ENH: Support arbitrary external files as completion sources via a:options.filespecs
Should a file be loaded as a Vim buffer, the buffer number is used, just like before. External files are loaded (and cached) via ingo#file#GetLines(). We use the file name in the menu then.
The a:alreadySearchedBuffers can handle arbitrary keys, so we can just put in the (absolute) filespec there, and it'll all work out.