Minor: Use new ingo#option#listchars#GetValues()
This enables the use of a three-character "tab" setting value in 'listchars'.
For processing speed (and because we don't need the full rendering of all 'listchars' settings, especially the expansion of the "tab" setting to a full tabstop width would be harmful for the length-limited completion popup), we keep our custom rendering here.